Fixes #10 by using the scrollbar expanded size
As described on issue #10 the linewrap plugin uses the docview width
minus current scrollbar width to calculate the amount of text that
can be visible. The scrollbar expand effect causes the wrapped text
flow to change everytime the scrollbar is hovered. This commit fixes
the issue by substracting from the docview width the full expanded
size of the scrollbar.