Better support the different supported runtimes
Currently LuaJIT and PUC Lua 5.4 has some runtime differences
that were not tackled by the pragtical_plugin_api.h which
provides to native plugins the ability to run using the same
Lua runtime that was linked to the editor.

This changes expose some buffer handling macros as functions for
both LuaJIT and PUC Lua to resolve this issue. It also exports the
LUA_REGISTRYINDEX macro as a "constant", since this value differs
between LuaJIT and PUC Lua 5.4 causing segmentation faults and
other related crashes.

The changes on this commit are an effort to support building
ltreesitter (needed by Evergreen plugin) using the plugin_api
mode, in order to support both LuaJIT and Lua 5.4 runtimes.