Added adaptive sizing to autocomplete plugin
This change handles the issue reported on lite-xl/lite-xl-lsp#82 about
the suggestions box becoming too wide to be visible when the symbols
text is too large.

Instead of delegating the responsability to plugins making use of
the autocomplete API we tackle the issue directly on the plugin
itself by performing the required calculations to resize the
suggestions box and truncate items text when needed.