autocomplete: strip punctuations from partials
When triggered manually and first character is a punctuation,
it causes none of the items to match if they don't also contain the
punctiation. We remove the punctuations to ensure results with plugins
like lsp will match even if the don't contain the starting punctuation
symbol.

This solves issues in languages like C/C++ where # causes lsp to return
a list of snippets as follows:

* if
* ifndef
* include

Without this change non of the items would match.