Add cli flag to allow forking editor to background
The new flag `-f` or `--fork` can be used to launch the editor
into a separate process when invoking it from a terminal. This
returns control back to the terminal wich reverse implements #131
but still tackles the same issue.

Currently explicit `--fork` is easier to implement and requires less
code than implicit forking and explicit `--wait`.