Return actual scale on get_scale() for macOS too
`system.get_scale()` always returned 1.0 (except the first time it
was called which was a bug :D), since older macOS builds always had
the renderer turned on and scaling was performed internally. That
is not the case anymore and the software renderer needs to know
current display scale when the displaychanged event is emitted.

This should fix the latest issue reported on #34 of UI elements
getting too small when switching from:

retina -> external display -> back to retina

Basically system.get_scale properly returned 2.0 the first time
the editor was started, then returned 1.0 when changing to external
display and 1.0 when returning back to the retina display causing
ui to get small.