Don't trust weak table keys to clean cached items
Various core plugins that implement some caching rely on weak table
keys of core.doc to cache content, with the idea that once the document
is closed the cached content will be cleaned. The reality is that
sometimes it can take a really long time until the garbage collector
decides that the key is no longer been referenced and the cache data
is kept longer than needed.

For this reason we now listen for Doc:on_close events and clean the cache
immediately, applicable plugins are:

* autocomplete
* autoreload
* detectindent
* drawwhitespace
* linewrapping

These changes introduce a new `core.collect_garbage` flag that
when set to true will queue a call to `collectgarbage('collect')`.
Currently it is been employed each time a document is closed in
an effort to reduce memory consumption.

We also remove the highlighter reference to the doc when closing
a doc to mitigate a circular reference issue.

Lastly, now the autoreload and detectindent plugins are applied to
visible documents only.