findfile: use project when resolving file path
When a single project is open we were wrongly using the currently
open file directory to calculate the absolute file path of selected
results entry even if currently open file do not belong to the project.

FindFile plugin goal is to open files related to currently open
projects, for extraneous files there is core:open-file.