Set coroutines max_time based on rendering speed
This changes tries to set the maximum execution time of co-routines
per frame iteration based on current frames per second.

It also penalizes co-routines that exceed the max_time by setting
their waiting time to the same time it took to execute them.

Also, since we are not keeping a record of the mean execution time
for coroutines, or the amount of time it takes for a core iteration,
we just reduce the max_time by a 50% to account for over-time scenarios.