highlighter: on long lines notify first/last only
Depending on installed plugins the update_notify call can get
expensive and slow things down. With this change, it is now perform
on first and last tokenization procedures, on first call it allows
the visible first characters of the line to be updated for correct
positioning calculations as plugins that need it.

The major drawback is when employing differently sized fonts for token
types. This may cause wrong positioning calculations when scrolling, but
it is preferred over a laggy code editor. It is not the norm to use
different font sizes for token types on files with long lines.
Currently only markdown language plugin makes such an use and users
that may have done heavy customization to the styling of the token types.