Lower input latency improving co-routine scheduling and tokenizer (#174)
* set the maximum execution time of co-routines - per frame iteration based on 
current frames per second.

* penalizes co-routines that exceed the max_time - by setting their waiting time to 
the same time it took to execute them.

* Check how much co-routines to run per frame - on each run_threads iteration we 
check how many can be run without exceeding the maximum time allowed.

* Use the avg. runtime of co-routines to yield - We now keep track of every 
1 second cycle in order to see how much time is left for executing co-routine tasks, 
by using the average execution time of co-routines to decide if we should execute 
it or skip it until next 1 second cycle. 

* introduce `core.fps`  - has the maximum frames per second that are actually 
possible to render, which can be used to accommodate to frame drops when 
performing some animations, scrolling, etc...

* tokenizer: always check time left on long lines - improves scrolling performance 
and other areas

* highlighter: on long lines notify first/last only - Depending on installed 
plugins the update_notify call can get expensive and slow things down. With this 
change, it is now perform on first and last tokenization procedures, on first call it 
allows the visible first characters of the line to be updated for correct positioning 
calculations as plugins that need it. The major drawback is when employing 
differently sized fonts for token types. This may cause wrong positioning calculations 
when scrolling, but it is preferred over a laggy code editor. It is not the norm to use 
different font sizes for token types on files with long lines. Currently only markdown 
language plugin makes such an use and users that may have done heavy customization 
to the styling of the token types.

* Expose co-routines max_time as core.co_max_time - Introduces 
core.co_max_time which is the maximum execution time for co-routines on a 
per frame iteration basis. Can be used by any task to determine how much time 
it has before yielding.

* Added new config.log_slow_threads option - This new configuration option 
allows monitoring of co-routines that are taking more time to execute than that 
allowed on a per frame iteration basis. Should help when troubleshooting 
performance issues.

* Main loop: burst events processing for 3s - When receiving events we 
reduce the wait time to match the rendering speed for the events that follow in 
order to process them earlier and reduce the input lag.