Restore DirWatch unwatch/watch behaviour
This change restores the unwatch/watch path behaviour when
receiving a change event which seemed to cause some issues on
the past. It is needed on backends like inotify because
when a file is deleted and re-added (like a git checkout <branch>)
the watch is lost internally so we need to readd it.