Restore DirWatch unwatch/watch behavior (#193)
This change restores the unwatch/watch path behavior when
receiving a change event which seemed to cause some issues on
the past. It is needed on backends like inotify because
when a file is deleted and re-added (like a git checkout <branch>)
the watch is lost internally so we need to re-add it.