Perform Lua unicode patching directly on Meson
On Windows PUC Lua needs a patch to properly handle paths
encoding. For sometime we already patched LuaJIT with meson
for the unicode support but, didn't yet for PUC Lua and
instead relied on the CI build scripts doing it for us
which causes issues for users that build pragtical on their own.