Disable a compatibility io overwrite on LuaJIT
For some context:

We use https://github.com/lunarmodules/lua-compat-5.3
to add some of the newer Lua 5.3+ changes and recently
we syncronized our copy with upstream on #224

The issue:

The compatibility code seems to have some workaround for
the io module when patching LuaJIT that causes the file
metatable type to be overwritten from `FILE*` to `userdata`,
causing issues with native functions that expect a `FILE*`.

Fixes #228, thanks to AmerM137 for reporting the issue.