Disable a compatibility io overwrite on LuaJIT (#229)
For some context:

We use https://github.com/lunarmodules/lua-compat-5.3
to add some of the newer Lua 5.3+ changes and recently
we synchronized our copy with upstream on #224

The issue:

The compatibility code seems to have some workaround for
the io module when patching LuaJIT that causes the file
metatable type to be overwritten from `FILE*` to `userdata`,
causing issues with native functions that expect a `FILE*`.
On a quick inspection, this workaround does not seems to 
be required.

Fixes #228, thanks to AmerM137 for reporting the issue.