Wait thread avg time when it exceeds max time (#264)
When a co-routine average time exceeded the maximum allowed
running time we skipped it for the incorrect time instead
of sleeping for its average time, this caused a delay in the
re-execution of subsequent tasks.

This change also recalculates the co-routines re-execution
wake time to account for the elpased time since the co-routine
with the highest run priority was executed.