Revert "don't fuzzy items without a needle (#269)"
This breaks the expected changes made to the table returned by
the fuzzy match function on various core and plugin components that
use it. The intended behavior of not sorting the commandview entries
should be handled on the commandview `suggest` function directly.
We already do this when no text is input on our version of find file.

This reverts commit 19db55f879034b38bc73e5dc1de328ac3519a0e6.