REPL history and completion support
This changes add linenoise which is the crossplatform counterpart
of readline and helps us provide a friendlier REPL with the
features you may have expected as back history support, completion of
built-in commands and as an added bonus completion of code symbols!

Also the REPL was ripped apart of core.cli to make it extensible by
plugins that may provide additional features not found out of the
box like additional commands or completion providers.

Support for linenoise can be disabled with the meson flag
`-Drepl_history=false`