Use LuaJIT FFI to perform draw calls (#288)
Use LuaJIT FFI to perform draw calls

With this change we bypass the Lua C API and instead call
the rencache drawing functions directly. The overhead of calling
into the drawing functions is reduced in around 90%, which is
a significant amount given that these functions are basically
the core rendering.

Some more overrides for frequently called functions:

* system.get_time
* system.sleep
* system.wait_event