Add special highlight colors to search selections (#312)
* Add special highlight colors to search selections

As noted by Amer and me included, most color schemes provide a
highlight selection color that is hard to spot when performing
search operations or selections as in this command`find-replace:select-add-all`.

For this reason this PR introduces the concept of a document search
selections. These search selections will be colored different than
a typical selection to make it more easy to identify. By default
they will use as background color the style.caret and for the text
the style.background. Two new style elements will be used if available:

* `style.search_selection` - background color of the selection
* `style.search_selection_text` - foreground (text) color of selection

When pressing the escape key (as with normal document selections) the
search selections will be cleared out.

* Adapted projectsearch to use add_search_selection

* Added an `instant` param to DocView:scroll_to_make_visible.

* Updated linewrapping plugin to respect the instant parameter on
override.

* Clear search selections on various docview events

* Make search_ui scroll to result instant

* Bonus: Add missing context line on bottom (as reported by Amer)