Revert back to performant iterator on each_token
Previously I added a feature to tokenizer.each_token where you
can give it the starting offset, this is used to more rapidly
calculate the x offset on the DocView. At that time I choose to
use a coroutine.wrap for the convenience but little I did know
about the performance difference between the previously used
function iterator and the coroutine.wrap + yield strategy.

In a small benchmark the difference of using the simpler outer
iterator function results on a 23 times speed increment from the
coroutine.wrap... so, this PR adapts the introduced starting offset
logic to an outer iterator function just as before.