Move inner functions to outer on tokenizer.tokenize (#318)
This can increase tokenization performance up to 50% or 2x as
shown on simple benchmark:

Inner functions       : 14.19575 sec (10000 iters)
Outer functions       : 7.27609 sec (10000 iters)

This is specially true for LuaJIT where it is able to better apply optimizations 
when tokenizer.tokenize is been deeply called. As an example, running the 
tokenizer with inner functions inside an if else condition will yield worst 
performance (as in the numbers above) than when run on the global scope.

Another benefit of moving the inner functions outside is a bit more of code clarity.