Add common.open_in_system(resource) function (#321)
* Add common.open_in_system(resource) function

This functionality was fragmented thru the source code as in
EmptyView, Settings and TreeView providing their own code to
open links, files and directories using the system launcher.

Also various plugins on the plugins repo provide their own way.
With this PR they can be adapted to use this function instead
of duplicating the functionality.

Besides providing a built-in mechanism of opening resources with
the operating system launcher, it also adds supports for other
operating systems:

* Amiga
* Haiku
* SerenityOS

Also the function should provide a central location for supporting
other systems if the need arises.

* Changes to code with walkero-gr guidance

Better handle AmigaOS since the Amiga port properly sets the
PLATFORM variable, also, properly handle URL's and directories.

On MorphOS the open command is available so use that.