Fixes performance regression by defer bitmap
As reported by Amer and discussed on #323. The real cause was the
code constantly trying to load the space character which is empty
by default, this ended in a lot of calling into freetype api causing
the performance issue.

Also, this PR caches the is_white_space check to reduce redundant
calls to it and re-utilizes font metrics on font_load_glyph_bitmap().