Account for missing bitmaps on glyph load or draw
Now, instead of skipping all white space we register if a glyph
is missing bitmap data and skip it. If not, we always draw it
even in the case of white space.

This is useful because some fonts like unifont provide glyph
representations on various white space codepoints that we can
use to easily show on a document, and in a more performant
way than with the drawwhitespace plugin.