treeview: fix a scale re-trigger issue
Previously, the on_scale_change event was introduced to properly
scale the treeview on dynamic scale changes, but the size config
on_apply was applying the user defined value already scaled which
was causing on_scale_change to enlarge the treeview even more.

This PR fixes that behaviour and always uses unscaled size on the
default and minimum values of the settings config.