Additional time based transition for scroll
This PR adds a new constant velocity transition with acceleration
support for use when performing scrolling on the DocView and
makes it the default for general transitions too. Also introduced
two new config flags:

* config.animation_type and
* config.scroll_animation_type

This lets the user choose the type of transition he wants to
"lerp" (previous) or "constant" (new and default).

The new constant transition as the name implies is more linear
and permits a smoother scroll movement which allows to better
read the text while scrolling. Also, when scrolling slowly the
DocView will be scrolled more smoothly, when scrolling rapidly
acceleration will be applied gradually to increment the speed,
the speed will decrease when reducing the scroll intensity.