Fix EmptyView AKA Welcome Screen flashing
This was caused by an out of order call of update and draw due
to how widgets bypass update and draw operations to reduce
cpu usage. The widget library was updated to prevent this issue
from happening.

Also this PR includes a change on settings ui to schedule an
update when the update method is called by the root system
without setting a width and height first, which is needed
to properly process the positioning calculations.