Fix system.list_dir to work with /
system.list_dir fails to return anything when '/' is given.

Not sure if this is a bug on SDL3 SDL_EnumerateDirectory function
but it seems that the internal path normalization functionality
strips a single '/' which is valid on posix OS like linux, mac, etc...

We replace '/' with '/.' to workaround the issue and get correct
directory listing.