Reorder current/default scales on get_display_info
On wayland SDL_GetDisplayContentScale() seems to always return
1, even if user changed the display scale factor. Meanwhile, the
display pixel_density does contains the proper scale factor.

On windows SDL_GetDisplayContentScale() properly returns the user
configured scale factor for the display while the display
pixel_density is set to 1.

Now we flip the values when this anomaly is detected to properly
return the current scale factor.

Hopefully this really fixes the issue reported by Amer on #393